+++ pkg/providers/cps/testdata/TestForceCertificateRenewalAction/missing_enrollment_id.tf
@@ -1,0 +1,18 @@
+provider "akamai" {
+  edgerc = "../../common/testutils/edgerc"
+}
+
+resource "terraform_data" "trigger" {
+  lifecycle {
+    action_trigger {
+      events  = [after_create]
+      actions = [action.akamai_cps_force_certificate_renewal.test]
+    }
+  }
+}
+
+action "akamai_cps_force_certificate_renewal" "test" {
+  config {
+    cancel_pending_changes = true
+  }
+}